+++ SongBeamerEdit/ValueConverter/FlagsEnumValueConverter.cs
@@ -6,21 +6,21 @@ namespace SongBeamerEdit.FlagsValueConverter
 {
     public class FlagsEnumValueConverter :IValueConverter
     {
-        private ushort targetValue;
+        private int targetValue;
 
         public FlagsEnumValueConverter() { }
 
         public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            ushort mask = (ushort)parameter;
-            this.targetValue = (ushort)value;
+            int mask = (int)parameter;
+            this.targetValue = (int)value;
 
             return ((mask & this.targetValue) != 0);
         }
 
         public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
         {
-            this.targetValue ^= (ushort)parameter;
+            this.targetValue ^= (int)parameter;
             return Enum.Parse(targetType, this.targetValue.ToString());
         }
     }
